LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/inside-dedicated-worker-expected.txt

CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8800/content-security-policy/support/ping.js because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.

Harness Error (TIMEOUT), message = null

PASS No SecurityPolicyViolation event fired for successful load.
TIMEOUT SecurityPolicyViolation event fired on global. Test timed out
TIMEOUT SecurityPolicyViolation event fired on global with the correct blockedURI. Test timed out
